requirements.txt
numpy
Pillow
scipy
torchvision
timm
torch
ultralytics
opencv-python-headless